Add Tests comparing the differences between prometheus and micrometer
These tests are too complex, but do exercise a fair bit of behaviour of both micometer and prometheus. The illustrate the differences that must be accounted for when rewriting prometheus metrics into micrometer metrics.

In the long run much simpler tests should be used, but for now these will get everything started.